+++ src/main/java/com/myclass/admin/controller/AdminTargetController.java
@@ -2,6 +2,7 @@ package com.myclass.admin.controller;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,15 +18,18 @@ import com.myclass.dto.TargetDto;
 import com.myclass.service.TargetService;
 
 /**
- * Author: Nguyen Chanh Truc
- * Created: Jan 29, 2021	
+ * Author: Nguyen Chanh Truc Created: Jan 29, 2021
  */
 @RestController
 @RequestMapping("api/admin/target")
 public class AdminTargetController {
-	
+	// Init message
+	@Value("${message.targetIdNotExist}")
+	private String targetIdNotExist;
+
 	// Inject TargetService
 	private TargetService targetService;
+
 	public AdminTargetController(TargetService targetService) {
 		this.targetService = targetService;
 	}
@@ -41,7 +45,7 @@ public class AdminTargetController {
 		}
 		return new ResponseEntity<Object>(HttpStatus.BAD_REQUEST);
 	}
-	
+
 	@GetMapping("{id}")
 	public Object get(@PathVariable int id) {
 		try {
@@ -50,11 +54,11 @@ public class AdminTargetController {
 			return new ResponseEntity<Object>(dto, HttpStatus.OK);
 		} catch (Exception e) {
 			e.printStackTrace();
-			return new ResponseEntity<Object>(e.getMessage(),HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<Object>(e.getMessage(), HttpStatus.BAD_REQUEST);
 		}
-		
+
 	}
-	
+
 	@PostMapping("")
 	public Object post(@RequestBody TargetDto dto) {
 		try {
@@ -62,37 +66,47 @@ public class AdminTargetController {
 			return new ResponseEntity<Object>(HttpStatus.CREATED);
 		} catch (Exception e) {
 			e.printStackTrace();
-			return new ResponseEntity<Object>(e.getMessage(),HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<Object>(e.getMessage(), HttpStatus.BAD_REQUEST);
 		}
-		
 	}
-	
+
 	@PutMapping("{id}")
 	public Object put(@PathVariable int id, @RequestBody TargetDto dto) {
 		try {
-			if(id != dto.getId()) {
+			// Check path id and targetId
+			if (id != dto.getId()) {
 				return new ResponseEntity<Object>(HttpStatus.BAD_REQUEST);
-			}		
+			}
+
+			// Check targetId exitence
+			if (!targetService.isTargetIdExist(id)) {
+				return new ResponseEntity<Object>(targetIdNotExist, HttpStatus.BAD_REQUEST);
+			}
+
 			targetService.edit(dto);
 			return new ResponseEntity<Object>(HttpStatus.OK);
-			
+
 		} catch (Exception e) {
 			e.printStackTrace();
-			return new ResponseEntity<Object>(e.getMessage(),HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<Object>(e.getMessage(), HttpStatus.BAD_REQUEST);
 		}
-		
+
 	}
 
 	@DeleteMapping("{id}")
 	public Object delete(@PathVariable int id) {
 		try {
+			// Check targetId exitence
+			if (!targetService.isTargetIdExist(id)) {
+				return new ResponseEntity<Object>(targetIdNotExist, HttpStatus.BAD_REQUEST);
+			}
+
 			targetService.delete(id);
 			return new ResponseEntity<Object>(HttpStatus.OK);
-			
+
 		} catch (Exception e) {
 			e.printStackTrace();
-			return new ResponseEntity<Object>(e.getMessage(),HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<Object>(e.getMessage(), HttpStatus.BAD_REQUEST);
 		}
-		
 	}
 }
